fix circular dependency, but really
