+++ editor/src/components/canvas/commands/adjust-css-length-command.ts
@@ -9,12 +9,11 @@ import {
   jsExpressionValue,
 } from '../../../core/shared/element-template'
 import type { ValueAtPath } from '../../../core/shared/jsx-attributes'
+import { setJSXValuesAtPaths, unsetJSXValuesAtPaths } from '../../../core/shared/jsx-attributes'
 import {
   getModifiableJSXAttributeAtPath,
   jsxSimpleAttributeToValue,
-  setJSXValuesAtPaths,
-  unsetJSXValuesAtPaths,
-} from '../../../core/shared/jsx-attributes'
+} from '../../../core/shared/jsx-attribute-utils'
 import { roundTo, roundToNearestWhole } from '../../../core/shared/math-utils'
 import type { ElementPath, PropertyPath } from '../../../core/shared/project-file-types'
 import * as PP from '../../../core/shared/property-path'
