+++ editor/src/components/canvas/commands/convert-to-absolute-command.spec.tsx
@@ -5,7 +5,7 @@ import * as EP from '../../../core/shared/element-path'
 import {
   getJSXAttributesAtPath,
   jsxSimpleAttributeToValue,
-} from '../../../core/shared/jsx-attributes'
+} from '../../../core/shared/jsx-attribute-utils'
 import { complexDefaultProjectPreParsed } from '../../../sample-projects/sample-project-utils.test-utils'
 import { withUnderlyingTargetFromEditorState } from '../../editor/store/editor-state'
 import { stylePropPathMappingFn } from '../../inspector/common/property-path-hooks'
