+++ editor/src/components/canvas/commands/set-css-length-command.ts
@@ -10,7 +10,7 @@ import type { GetModifiableAttributeResult, ValueAtPath } from '../../../core/sh
 import {
   getModifiableJSXAttributeAtPath,
   jsxSimpleAttributeToValue,
-} from '../../../core/shared/jsx-attributes'
+} from '../../../core/shared/jsx-attribute-utils'
 import { roundTo } from '../../../core/shared/math-utils'
 import type { ElementPath, PropertyPath } from '../../../core/shared/project-file-types'
 import * as PP from '../../../core/shared/property-path'
