+++ editor/src/components/editor/actions/actions.spec.tsx
@@ -40,10 +40,8 @@ import {
   unparsedCode,
   utopiaJSXComponent,
 } from '../../../core/shared/element-template'
-import {
-  clearModifiableAttributeUniqueIDs,
-  getModifiableJSXAttributeAtPath,
-} from '../../../core/shared/jsx-attributes'
+import { clearModifiableAttributeUniqueIDs } from '../../../core/shared/jsx-attributes'
+import { getModifiableJSXAttributeAtPath } from '../../../core/shared/jsx-attribute-utils'
 import type { CanvasRectangle, LocalRectangle } from '../../../core/shared/math-utils'
 import { canvasRectangle, zeroRectangle } from '../../../core/shared/math-utils'
 import { resolvedNpmDependency } from '../../../core/shared/npm-dependency-types'
