+++ editor/src/components/editor/insert-callbacks.ts
@@ -39,7 +39,7 @@ import { assertNever } from '../../core/shared/utils'
 import type { ComponentElementToInsert } from '../custom-code/code-file'
 import { notice } from '../common/notice'
 import * as PP from '../../core/shared/property-path'
-import { setJSXValueInAttributeAtPath } from '../../core/shared/jsx-attributes'
+import { setJSXValueInAttributeAtPath } from '../../core/shared/jsx-attribute-utils'
 import { defaultEither, isLeft } from '../../core/shared/either'
 import { executeFirstApplicableStrategy } from '../inspector/inspector-strategies/inspector-strategy'
 import { insertAsAbsoluteStrategy } from './one-shot-insertion-strategies/insert-as-absolute-strategy'
