+++ editor/src/components/inspector/common/inspector.test-utils.tsx
@@ -4,7 +4,7 @@ import create from 'zustand'
 import { subscribeWithSelector } from 'zustand/middleware'
 import type { JSXElement } from '../../../core/shared/element-template'
 import { emptyComments, jsExpressionValue } from '../../../core/shared/element-template'
-import { setJSXValueAtPath } from '../../../core/shared/jsx-attributes'
+import { setJSXValueAtPath } from '../../../core/shared/jsx-attribute-utils'
 import { isRight } from '../../../core/shared/either'
 import type {
   PropertyPath,
