+++ editor/src/components/inspector/sections/component-section/component-section.tsx
@@ -100,11 +100,6 @@ import {
   jsIdentifier,
 } from '../../../../core/shared/element-template'
 import { optionalMap } from '../../../../core/shared/optional-utils'
-import {
-  ModifiableAttribute,
-  getJSExpressionAtPath,
-  getJSXAttributesAtPath,
-} from '../../../../core/shared/jsx-attributes'
 import type { VariableData } from '../../../canvas/ui-jsx-canvas'
 import { array } from 'prop-types'
 import { useVariablesInScopeForSelectedElement } from './variables-in-scope-utils'
