+++ editor/src/components/text-editor/text-handling.ts
@@ -19,7 +19,7 @@ import {
   isJSXAttributesEntry,
   isJSXConditionalExpression,
 } from '../../core/shared/element-template'
-import { jsxSimpleAttributeToValue } from '../../core/shared/jsx-attributes'
+import { jsxSimpleAttributeToValue } from '../../core/shared/jsx-attribute-utils'
 import { foldEither } from '../../core/shared/either'
 import fastDeepEquals from 'fast-deep-equal'
 import { getUtopiaID } from '../../core/shared/uid-utils'
