+++ editor/src/core/model/jsx-attributes.spec.ts
@@ -28,14 +28,16 @@ import {
   clearAttributesUniqueIDs,
 } from '../shared/element-template'
 import {
-  dropKeyFromNestedObject,
   getAllPathsFromAttributes,
-  getModifiableJSXAttributeAtPath,
   jsxAttributesToProps,
-  jsxSimpleAttributeToValue,
-  setJSXValueAtPath,
   unsetJSXValueAtPath,
 } from '../shared/jsx-attributes'
+import {
+  dropKeyFromNestedObject,
+  getModifiableJSXAttributeAtPath,
+  jsxSimpleAttributeToValue,
+  setJSXValueAtPath,
+} from '../shared/jsx-attribute-utils'
 import { Substores } from '../../components/editor/store/store-hook'
 import { emptyUiJsxCanvasContextData } from '../../components/canvas/ui-jsx-canvas'
 import { NO_OP } from '../shared/utils'
