+++ editor/src/core/model/scene-id-utils.ts
@@ -5,7 +5,7 @@ import { StoryboardFilePath } from '../../components/editor/store/editor-state'
 import { isLeft } from '../shared/either'
 import type { JSXElementChild, ElementsWithin, JSXElement } from '../shared/element-template'
 import { getJSXAttribute, jsExpressionValue, emptyComments } from '../shared/element-template'
-import { setJSXValueAtPath } from '../shared/jsx-attributes'
+import { setJSXValueAtPath } from '../shared/jsx-attribute-utils'
 import type { TextFile } from '../shared/project-file-types'
 import * as PP from '../shared/property-path'
 import { assertNever } from '../shared/utils'
